# feat: introduce new event listener abstraction