--- a/foyer/src/hybrid/builder.rs
+++ b/foyer/src/hybrid/builder.rs
@@ -15,7 +15,10 @@
 use std::{fmt::Debug, sync::Arc};
 
 use ahash::RandomState;
-use foyer_common::code::{HashBuilder, StorageKey, StorageValue};
+use foyer_common::{
+    code::{HashBuilder, StorageKey, StorageValue},
+    event::EventListener,
+};
 use foyer_memory::{Cache, CacheBuilder, EvictionConfig, Weighter};
 use foyer_storage::{
     AdmissionPicker, Compression, DeviceConfig, EvictionPicker, RecoverMode, ReinsertionPicker, RuntimeConfig,
@@ -25,21 +28,23 @@ use foyer_storage::{
 use crate::HybridCache;
 
 /// Hybrid cache builder.
-pub struct HybridCacheBuilder {
+pub struct HybridCacheBuilder<K, V> {
     name: String,
+    event_listener: Option<Arc<dyn EventListener<Key = K, Value = V>>>,
 }
 
-impl Default for HybridCacheBuilder {
+impl<K, V> Default for HybridCacheBuilder<K, V> {
     fn default() -> Self {
         Self::new()
     }
 }
 
-impl HybridCacheBuilder {
+impl<K, V> HybridCacheBuilder<K, V> {
     /// Create a new hybrid cache builder.
     pub fn new() -> Self {
         Self {
             name: "foyer".to_string(),
+            event_listener: None,
         }
     }
 
@@ -53,14 +58,26 @@ impl HybridCacheBuilder {
         self
     }
 
+    /// Set event listener.
+    ///
+    /// Default: No event listener installed.
+    pub fn with_event_listener(mut self, event_listener: Arc<dyn EventListener<Key = K, Value = V>>) -> Self {
+        self.event_listener = Some(event_listener);
+        self
+    }
+
     /// Continue to modify the in-memory cache configurations.
-    pub fn memory<K, V>(self, capacity: usize) -> HybridCacheBuilderPhaseMemory<K, V, RandomState>
+    pub fn memory(self, capacity: usize) -> HybridCacheBuilderPhaseMemory<K, V, RandomState>
     where
         K: StorageKey,
         V: StorageValue,
     {
+        let mut builder = CacheBuilder::new(capacity).with_name(&self.name);
+        if let Some(event_listener) = self.event_listener {
+            builder = builder.with_event_listener(event_listener);
+        }
         HybridCacheBuilderPhaseMemory {
-            builder: CacheBuilder::new(capacity).with_name(&self.name),
+            builder,
             name: self.name,
         }
     }