@MrCroxx 2024-06-03
## Benchmark

```bash
rm -rf /6536/foyer && cargo build --release  && RUST_BACKTRACE=1 RUST_LOG=info ./target/release/foyer-bench --dir /6536/foyer --mem 64 --disk 102400 --file-size 64 --get-range 10000 --flushers 4 --reclaimers 4 --time 60 --writers 256 --w-rate 4 --admission-rate-limit 500 --readers 1024 --r-rate 1 --runtime --compression none --distribution zipf --distribution-zipf-s 0.8 --metrics --eviction lru
```

### main (7f6b5dda44d8bfa4de7be136a0c01470601df825)

```
Total:
disk total iops: 2731.2
disk total throughput: 509.4 MiB/s
disk read iops: 126.7
disk read throughput: 8.4 MiB/s
disk write iops: 2604.5
disk write throughput: 501.0 MiB/s
insert iops: 14283.5/s
insert throughput: 892.8 MiB/s
insert lat p50: 5us
insert lat p90: 9us
insert lat p99: 21us
insert lat p999: 76us
insert lat p9999: 455us
insert lat p99999: 619us
insert lat pmax: 1135us
get iops: 30312.6/s
get miss: 47.74%
get throughput: 1.0 GiB/s
get hit lat p50: 3us
get hit lat p90: 459us
get hit lat p99: 1303us
get hit lat p999: 2431us
get hit lat p9999: 5247us
get hit lat p99999: 5439us
get hit lat pmax: 5503us
get miss lat p50: 158us
get miss lat p90: 547us
get miss lat p99: 1239us
get miss lat p999: 2303us
get miss lat p9999: 3791us
get miss lat p99999: 4735us
get miss lat pmax: 5247us
```

### PR

```
Total:
disk total iops: 2755.3
disk total throughput: 510.2 MiB/s
disk read iops: 140.8
disk read throughput: 9.3 MiB/s
disk write iops: 2614.5
disk write throughput: 500.9 MiB/s
insert iops: 14283.9/s
insert throughput: 892.9 MiB/s
insert lat p50: 5us
insert lat p90: 9us
insert lat p99: 21us
insert lat p999: 67us
insert lat p9999: 469us
insert lat p99999: 635us
insert lat pmax: 739us
get iops: 30039.2/s
get miss: 47.27%
get throughput: 1.0 GiB/s
get hit lat p50: 1us
get hit lat p90: 427us
get hit lat p99: 1319us
get hit lat p999: 2415us
get hit lat p9999: 5055us
get hit lat p99999: 5215us
get hit lat pmax: 5247us
get miss lat p50: 107us
get miss lat p90: 413us
get miss lat p99: 1143us
get miss lat p999: 2975us
get miss lat p9999: 6495us
get miss lat p99999: 8511us
get miss lat pmax: 8575us
```